+++ ReminderBox/clock.h
@@ -1,10 +1,12 @@
 #ifndef CLOCK_H_
 #define CLOCK_H_
 
+#include <stdint.h>
+
 void clock_init(void);
-void clock_delay_ms(unsigned int to_delay);
-unsigned long clock_get_ms(void);
-unsigned long clock_get_seconds(void);
+void clock_delay_ms(uint32_t to_delay);
+uint32_t clock_get_ms(void);
+uint32_t clock_get_seconds(void);
 void clock_start_timer(void);
